# Using uv_buf_init on the allocation callback, to ensure platform-inde…