@kellabyte 2016-01-15
I think we need to also call `uv_buf_init()` here:
https://github.com/haywire/haywire/blob/master/src/haywire/http_server.c#L281