@nmdguerreiro 2016-01-19
@kellabyte can this be merged?