@nmdguerreiro 2016-01-19
Closing as this will be reflected in my newest PR.